create location service in the API for location functions
