+++ api/src/features/growRunLocation.ts
@@ -1,20 +1,7 @@
-import type { Coords } from '@grow-run-archive/definitions';
 import { Request, Response, Router } from 'express';
 import { StatusCodes } from 'http-status-codes';
 import { EntityController } from '@/entity/controller.js';
-
-export async function reverseGeocode(latitude: Coords['latitude'], longitude: Coords['longitude']) {
-	const locationIQAPIUrl =
-		`https://us1.locationiq.com/v1/reverse?` +
-		new URLSearchParams({
-			lat: latitude.toString(),
-			lon: longitude.toString(),
-			format: 'json',
-			key: process.env.SECRET_LOCATION_IQ_API_KEY as string
-		});
-
-	return await (await fetch(locationIQAPIUrl)).json();
-}
+import { reverseGeocode } from '@/services/location/locationIQ.js';
 
 class Controller implements EntityController {
 	async create(req: Request, res: Response) {
